+++ src/adapter/types.ts
@@ -1,0 +1,15 @@
+export type EventPublish = {
+  appId: string;
+  channel: string;
+  event: string;
+  data: string;
+  excludeSocket?: string;
+  userId?: string;
+};
+
+export interface EventAdapter {
+  initialize(): Promise<void>;
+  publish(event: EventPublish): Promise<boolean>;
+  receive(event: EventPublish): boolean;
+  close(): Promise<void>;
+}